implement tryGet for all funs

+++ R/emuia_download.R
@@ -58,7 +58,12 @@ emuia_download = function(commune = NULL, TERYT = NULL, outdir = ".",
 
   for (i in seq_len(nrow(df_names))) {
     filename = paste0(outdir, "/", df_names[i, "TERYT"], ".zip")
-    utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...)
+    status = tryGet(utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...))
+
+    if (status %in% c("error", "warning")) {
+      return("connection error")
+    }
+
     if (unzip) {
       utils::unzip(filename, exdir = outdir)
       file.remove(filename)
