+++ R/geonames_download.R
@@ -52,7 +52,12 @@ geonames_download = function(type, outdir = ".", unzip = TRUE, format = "SHP", .
   for (i in seq_len(nrow(df))) {
     filename = paste0(outdir, "/", df[i, "type"], "_",
                       df[i, "format"], ".zip")
-    utils::download.file(df[i, "URL"], filename, mode = "wb", ...)
+    status = tryGet(utils::download.file(df[i, "URL"], filename, mode = "wb", ...))
+
+    if (status %in% c("error", "warning")) {
+      return("connection error")
+    }
+
     if (unzip) {
       utils::unzip(filename, exdir = outdir)
       file.remove(filename)
